Use Case Projects/19 mail send 🦢/mail2/requirements.txt
Flask
pandas
openpyxl
python-dotenv
